Implement review comments for DefaultWallet changes

+++ browser/brave_prefs_browsertest.cc
@@ -32,6 +32,7 @@
 
 #if BUILDFLAG(BRAVE_WALLET_ENABLED)
 #include "brave/components/brave_wallet/browser/brave_wallet_utils.h"
+#include "brave/components/brave_wallet/browser/pref_names.h"
 #include "brave/components/brave_wallet/common/brave_wallet.mojom.h"
 #endif
 
@@ -109,12 +110,11 @@ IN_PROC_BROWSER_TEST_F(BraveProfilePrefsBrowserTest, MiscBravePrefs) {
       kERCOptedIntoCryptoWallets));
 #endif
 #if BUILDFLAG(BRAVE_WALLET_ENABLED)
-  EXPECT_EQ(
-      brave_wallet::GetDefaultWallet(
-          chrome_test_utils::GetProfile(this)->GetPrefs()),
-      static_cast<int>(brave_wallet::IsNativeWalletEnabled()
-                           ? brave_wallet::mojom::DefaultWallet::BraveWallet
-                           : brave_wallet::mojom::DefaultWallet::Ask));
+  EXPECT_EQ(brave_wallet::GetDefaultWallet(
+                chrome_test_utils::GetProfile(this)->GetPrefs()),
+            brave_wallet::IsNativeWalletEnabled()
+                ? brave_wallet::mojom::DefaultWallet::BraveWallet
+                : brave_wallet::mojom::DefaultWallet::Ask);
   EXPECT_TRUE(chrome_test_utils::GetProfile(this)->GetPrefs()->GetBoolean(
       kShowWalletIconOnToolbar));
   EXPECT_FALSE(chrome_test_utils::GetProfile(this)->GetPrefs()->GetBoolean(
